--- client/src/app/controllers/add-edit-assessment/add-edit-assessment.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AddEditAssessmentService } from './add-edit-assessment.service';
-
-describe('AddEditAssessmentService', () => {
-  let service: AddEditAssessmentService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AddEditAssessmentService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
