--- client/src/app/controllers/add-edit-assessment/add-edit-assessment.service.ts
@@ -1,19 +1,0 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { AuthService } from '../auth-controller/auth.service';
-import { Observable } from 'rxjs';
-import { environment } from 'src/environments/environment';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AddEditAssessmentService {
-  private apiUrl = environment.apiUrl;
-
-  constructor(private http: HttpClient, private authService: AuthService) { }
-
-  createAssessment(data: any): Observable<any> {
-    const headers = this.authService.getHeaders();
-    return this.http.post(`${this.apiUrl}/industry/createAssessment`, data, { headers });
-  }
-}
